+++ android/src/main/java/com/transistorsoft/rnbackgroundfetch/RNBackgroundFetchModule.java
@@ -3,6 +3,8 @@ package com.transistorsoft.rnbackgroundfetch;
 import android.app.Activity;
 import android.content.Intent;
 
+import androidx.annotation.Nullable;
+
 import com.facebook.react.bridge.*;
 import com.facebook.react.modules.core.RCTNativeAppEventEmitter;
 import com.transistorsoft.tsbackgroundfetch.BackgroundFetch;
@@ -12,7 +14,9 @@ public class RNBackgroundFetchModule extends ReactContextBaseJavaModule implemen
     public static final String TAG = "RNBackgroundFetch";
 
     private static final String EVENT_FETCH = "fetch";
-    private static final String JOB_SERVICE_CLASS = "HeadlessJobService";
+    private static final String JOB_SERVICE_CLASS = HeadlessTask.class.getName();
+    private static final String FETCH_TASK_ID                       = "react-native-background-fetch";
+
     private boolean isForceReload = false;
     private boolean initialized = false;
 
@@ -22,6 +26,8 @@ public class RNBackgroundFetchModule extends ReactContextBaseJavaModule implemen
     }
 
     @Override
+
+
     public String getName() {
         return TAG;
     }
@@ -30,63 +36,37 @@ public class RNBackgroundFetchModule extends ReactContextBaseJavaModule implemen
     public void configure(ReadableMap options, final Callback failure) {
         BackgroundFetch adapter = getAdapter();
 
-        BackgroundFetchConfig.Builder config = new BackgroundFetchConfig.Builder();
-        if (options.hasKey("minimumFetchInterval")) {
-            config.setMinimumFetchInterval(options.getInt("minimumFetchInterval"));
-        }
-        if (options.hasKey("stopOnTerminate")) {
-            config.setStopOnTerminate(options.getBoolean("stopOnTerminate"));
-        }
-        if (options.hasKey("forceReload")) {
-            config.setForceReload(options.getBoolean("forceReload"));
-        }
-        if (options.hasKey("startOnBoot")) {
-            config.setStartOnBoot(options.getBoolean("startOnBoot"));
-        }
-        if (options.hasKey("enableHeadless") && options.getBoolean("enableHeadless")) {
-            config.setJobService(getClass().getPackage().getName() + "." + JOB_SERVICE_CLASS);
-        }
-        if (options.hasKey("requiredNetworkType")) {
-            config.setRequiredNetworkType(options.getInt("requiredNetworkType"));
-        }
-        if (options.hasKey("requiresBatteryNotLow")) {
-            config.setRequiresBatteryNotLow(options.getBoolean("requiresBatteryNotLow"));
-        }
-        if (options.hasKey("requiresCharging")) {
-            config.setRequiresCharging(options.getBoolean("requiresCharging"));
-        }
-        if (options.hasKey("requiresDeviceIdle")) {
-            config.setRequiresDeviceIdle(options.getBoolean("requiresDeviceIdle"));
-        }
-        if (options.hasKey("requiresStorageNotLow")) {
-            config.setRequiresStorageNotLow(options.getBoolean("requiresStorageNotLow"));
-        }
-
         BackgroundFetch.Callback callback = new BackgroundFetch.Callback() {
-            @Override
-            public void onFetch() {
-                WritableMap params = new WritableNativeMap();
-                getReactApplicationContext().getJSModule(RCTNativeAppEventEmitter.class).emit(EVENT_FETCH, params);
+            @Override public void onFetch(String taskId) {
+                getReactApplicationContext().getJSModule(RCTNativeAppEventEmitter.class).emit(EVENT_FETCH, taskId);
             }
         };
-        adapter.configure(config.build(), callback);
-        if (isForceReload) {
-            callback.onFetch();
-        }
-        isForceReload = false;
+        adapter.configure(buildConfig(options)
+                .setTaskId(FETCH_TASK_ID)
+                .setIsFetchTask(true)
+                .build(), callback);
+
+    }
+
+    @ReactMethod
+    public void scheduleTask(ReadableMap options, final Callback success, final Callback failure) {
+        BackgroundFetch adapter = getAdapter();
+        adapter.scheduleTask(buildConfig(options).build());
+        success.invoke(true);
     }
 
     @ReactMethod
     public void start(Callback success, Callback failure) {
         BackgroundFetch adapter = getAdapter();
-        adapter.start();
+        adapter.start(FETCH_TASK_ID);
         success.invoke(adapter.status());
     }
 
     @ReactMethod
-    public void stop() {
+    public void stop(@Nullable String taskId) {
+        if (taskId == null) taskId = FETCH_TASK_ID;
         BackgroundFetch adapter = getAdapter();
-        adapter.stop();
+        adapter.stop(taskId);
     }
 
     @ReactMethod
@@ -96,9 +76,9 @@ public class RNBackgroundFetchModule extends ReactContextBaseJavaModule implemen
     }
 
     @ReactMethod
-    public void finish(Integer fetchResult) {
+    public void finish(String taskId) {
         BackgroundFetch adapter = getAdapter();
-        adapter.finish();
+        adapter.finish(taskId);
     }
 
     @Override
@@ -126,17 +106,52 @@ public class RNBackgroundFetchModule extends ReactContextBaseJavaModule implemen
 
     }
 
+    private BackgroundFetchConfig.Builder buildConfig(ReadableMap options) {
+        BackgroundFetchConfig.Builder config = new BackgroundFetchConfig.Builder();
+        if (options.hasKey(BackgroundFetchConfig.FIELD_MINIMUM_FETCH_INTERVAL)) {
+            config.setMinimumFetchInterval(options.getInt(BackgroundFetchConfig.FIELD_MINIMUM_FETCH_INTERVAL));
+        }
+        if (options.hasKey(BackgroundFetchConfig.FIELD_TASK_ID)) {
+            config.setTaskId(options.getString(BackgroundFetchConfig.FIELD_TASK_ID));
+        }
+        if (options.hasKey(BackgroundFetchConfig.FIELD_DELAY)) {
+            Integer delay = options.getInt(BackgroundFetchConfig.FIELD_DELAY);
+            config.setDelay(delay.longValue());
+        }
+        if (options.hasKey(BackgroundFetchConfig.FIELD_STOP_ON_TERMINATE)) {
+            config.setStopOnTerminate(options.getBoolean(BackgroundFetchConfig.FIELD_STOP_ON_TERMINATE));
+        }
+        if (options.hasKey(BackgroundFetchConfig.FIELD_FORCE_ALARM_MANAGER)) {
+            config.setForceAlarmManager(options.getBoolean(BackgroundFetchConfig.FIELD_FORCE_ALARM_MANAGER));
+        }
+        if (options.hasKey(BackgroundFetchConfig.FIELD_START_ON_BOOT)) {
+            config.setStartOnBoot(options.getBoolean(BackgroundFetchConfig.FIELD_START_ON_BOOT));
+        }
+        if (options.hasKey("enableHeadless") && options.getBoolean("enableHeadless")) {
+            config.setJobService(JOB_SERVICE_CLASS);
+        }
+        if (options.hasKey(BackgroundFetchConfig.FIELD_REQUIRED_NETWORK_TYPE)) {
+            config.setRequiredNetworkType(options.getInt(BackgroundFetchConfig.FIELD_REQUIRED_NETWORK_TYPE));
+        }
+        if (options.hasKey(BackgroundFetchConfig.FIELD_REQUIRES_BATTERY_NOT_LOW)) {
+            config.setRequiresBatteryNotLow(options.getBoolean(BackgroundFetchConfig.FIELD_REQUIRES_BATTERY_NOT_LOW));
+        }
+        if (options.hasKey(BackgroundFetchConfig.FIELD_REQUIRES_CHARGING)) {
+            config.setRequiresCharging(options.getBoolean(BackgroundFetchConfig.FIELD_REQUIRES_CHARGING));
+        }
+        if (options.hasKey(BackgroundFetchConfig.FIELD_REQUIRES_DEVICE_IDLE)) {
+            config.setRequiresDeviceIdle(options.getBoolean(BackgroundFetchConfig.FIELD_REQUIRES_DEVICE_IDLE));
+        }
+        if (options.hasKey(BackgroundFetchConfig.FIELD_REQUIRES_STORAGE_NOT_LOW)) {
+            config.setRequiresStorageNotLow(options.getBoolean(BackgroundFetchConfig.FIELD_REQUIRES_STORAGE_NOT_LOW));
+        }
+        return config;
+    }
+
     private void initializeBackgroundFetch() {
         Activity activity = getCurrentActivity();
         if (activity == null) {
             return;
-        }
-        Intent intent = activity.getIntent();
-        String action = intent.getAction();
-
-        if ((action != null) && (BackgroundFetch.ACTION_FORCE_RELOAD.equalsIgnoreCase(action))) {
-            isForceReload = true;
-            activity.moveTaskToBack(true);
         }
         initialized = true;
     }